model/requirements.txt
concrete-ml
scikit-learn
pandas
numpy
matplotlib
msgpack
requests
click
grequests
web3